Add TreeViewData validator, move types and reducer in separate files, fix review notes.
