+++ README.md
@@ -41,7 +41,7 @@ A react component that implements the treeview pattern as described by the [WAI-
 
 ## data
 
-Data can be eigther array of nodes or Map of nodes with unique and non-sequential ids.
+Data can be either array of nodes or Map of nodes with unique and non-sequential ids.
 
 Nodes are objects with the following structure:
