+++ website/docs/examples/ControlledExpandedNode/index.js
@@ -62,7 +62,12 @@ function ControlledExpandedNode() {
       document
         .querySelector("#txtIdsToExpand")
         .value.split(",")
-        .map((x) => parseInt(x.trim()))
+        .map((x) => {
+          if (isNaN(parseInt(x.trim()))) {
+            return x;
+          }
+          return parseInt(x.trim());
+        })
     );
   };
